Revert to flattened components folder. Move FormText.tsx into 'shared' folder. Add better comments to FormText components
