+++ packages/components/src/index.tsx
@@ -9,8 +9,8 @@ if (expoApp && App.initiateExpo) {
 // Export components here so they are exported through npm
 export { Alert } from './components/Alert/Alert'
 export { Button, ButtonVariants } from './components/Button/Button'
-export { Checkbox } from './components/Forms/Checkbox/Checkbox'
-export { CheckboxGroup } from './components/Forms/CheckboxGroup/CheckboxGroup'
+export { Checkbox } from './components/Checkbox/Checkbox'
+export { CheckboxGroup } from './components/CheckboxGroup/CheckboxGroup'
 export { Icon } from './components/Icon/Icon'
 export { Link } from './components/Link/Link'
 export { LoadingIndicator } from './components/LoadingIndicator/LoadingIndicator'
